requirements.txt
click==8.0.3
colorama==0.4.4
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.2
Flask-MySQL==1.5.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.28
protobuf==3.19.4
PyMySQL==1.0.2
SQLAlchemy==1.4.31
Werkzeug==2.0.2
WTForms==3.0.1
